+++ framework/app/app.gotext
@@ -56,7 +56,7 @@ func (a *App) Run(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	budClient, err := budclient.Try(os.Getenv("BUD_LISTEN"))
+	budClient, err := budhttp.Try(os.Getenv("BUD_LISTEN"))
 	if err != nil {
 		return err
 	}
@@ -78,7 +78,7 @@ func (a *App) Run(ctx context.Context) error {
 	webServer, err := loadWeb(
 		{{/* Order matters. Ordered by package name */}}
 		{{- if $.Provider.Variable "context.Context" }}ctx,{{ end }}
-		{{- if $.Provider.Variable "github.com/livebud/bud/package/budclient.Client" }}budClient,{{ end }}
+		{{- if $.Provider.Variable "github.com/livebud/bud/package/budhttp.Client" }}budClient,{{ end }}
 		{{- if $.Provider.Variable "github.com/livebud/bud/package/gomod.*Module" }}module,{{ end }}
 		{{- if $.Provider.Variable "github.com/livebud/bud/package/log.Interface" }}log,{{ end }}
 	)
